Successfully generate Enum header files.
Added df_common module that defines the SharedState class that stores variables needed outside of the main df_common and also various functions that act on these variables.
df_common also has a few static functions and constant variables.

Added df_enum module that defines the EnumWrapper object. EnumWrapper takes an enum's root etree._Element node and a shared_state instance and generates most of the parameters needed to be passed to the enum.h.jinja template.

enum.h.jinja now includes comments and headers that are needed by the enum-attr attributes.
The comment printing logic is used in two places and thus should be a jinja macro, but my attempt at moving it to a macro caused the whitespace to be messed up in the generated header files, so it is copied and pasted for now.